+++ platform_orchestrator/src/main/java/com/override/controller/rest/QuestionRestController.java
@@ -4,7 +4,9 @@ import com.override.model.Question;
 import com.override.service.CustomStudentDetailService;
 import com.override.service.LessonStructureService;
 import com.override.service.QuestionService;
+import dto.QuestionDTO;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
@@ -34,5 +36,23 @@ public class QuestionRestController {
     @GetMapping("/chapters")
     public List<String> findAllChapters() {
         return lessonStructureService.getChapterNamesList();
+    }
+
+    @Secured("ROLE_ADMIN")
+    @PostMapping
+    public void save(@RequestBody QuestionDTO questionDTO) {
+        questionService.save(questionDTO);
+    }
+
+    @Secured("ROLE_ADMIN")
+    @PatchMapping
+    public void patch(@RequestBody QuestionDTO questionDTO) {
+        questionService.patch(questionDTO);
+    }
+
+    @Secured("ROLE_ADMIN")
+    @DeleteMapping
+    public void delete(@RequestParam long id) {
+        questionService.delete(id);
     }
 }
